root files don't need to be split
When we detect CSS files that contain `@utility` and non-`@utility`
parts, then we split them. We will also split every parent file to
ensure the imports are correct.

The new files with `@utility` should not result in a `layer(…)` next to
the imports. The other CSS might end up with a `layer(…)`