add `memcpy` util
It's a little bit silly, but during debugging I noticed that the
`variant` looked all funky. This is because we convert one variant to
another via `Object.assign`.

It does work properly because we compare the `kind` field (which gets
converted), and then we know that all the properties of the new `kind`
are available.

With this change we first clear out all the existing keys, resulting in
an expected shape when inspecting the `variant`.