expose the type of the CSS file
It could be that the Tailwind compiler was invoked with a CSS function
that doesn't use Tailwind at all. This can happen in places like the
PostCSS or Vite plugin where _every_ CSS file is processed.

The type of the CSS file can be any of 3 types. The exact naming is to
be determined. But the idea is:

1. `none` — nothing related to Tailwind CSS is being used. This means
   that we can skip handling this file.
2. `static` — Some Tailwind CSS features are used such as `@apply` or
   the `theme(…)` function. Tailwind still has to handle these
   situations. The important part here is that `@tailwind utilities` was
   not used. This means that we don't even have to scan for candidates
   at all.
4. `full` — `@tailwind utilities` was used, this means that we have to
   scan for candidates and do a lot more I/O.