run the extractor on very small chunks, split by whitespace
We already know that we can't have whitespace in our candidates which
means that we can operate on many many small chunks instead.

This change means that we run a lot of extractors on very small pieces
of code all in parallel, instead of running the extractor per file.

In my testing, this results in 2x speedup (~800ms -> ~400ms on a big
project with over a 1000 files).