properly wait for processes to be ready
This is the biggest change. Whenever we start a long running process
(via `spawn(…)`), we will ensure to wait for specific message on
`stdout` or `stderr` from the process.

These typically indicate when a process is ready (e.g.: initial build is
done, or watchers are setup and we are waiting for file changes).

This fixes race conditions where we already start writing files and
expect a rebuild to happen before the process is ready. These messages
can differ per command we run, so they are implemented on a per-test
basis.

Additionally, most long running processes that require a dev server
expose the URL (host + port) in the stdout/stderr. We will use that
information instead of using the `getFreePort()` that we used to use.

The big benefit is that these processes typically already have conflict
erroring, and try until a free port is available.

Because most tests are concurrent, we also use the `expect` exposed from
the test helper.