Only compile arbitrary values ending in `]` (#15503)
This PR ensures that if you are using an arbitrary value such as
`bg-[red` that it only compiles if it's a properly closed arbitrary
value.

Currently what happens is that it assumes the `]` is there, and cuts it
off. This then results in the `bg-[red` to be compiled as:

```css
.bg-\[red {
  background-color: re;
}
```

Note how the `d` in `red` is cut off. That's the assumption that the `]`
is there.

This PR fixes that by ensuring that the arbitrary value is properly
closed.

Fixes: #15484