update error message
A circular dependency means that we are applying in a circle, so while
this changed it's just the _other_ side of the circle (are there even
sides on a circle? Help?)

In these cases, the errors are technically better, because this is the
first node we see where the circle starts.